+++ src/logger.ts
@@ -1,71 +1,77 @@
 'use strict';
-import { ExtensionContext, OutputChannel, window, workspace } from 'vscode';
-import { IConfig } from './configuration';
-import { ExtensionKey, ExtensionOutputChannelName } from './constants';
+import { ConfigurationChangeEvent, ExtensionContext, OutputChannel, window } from 'vscode';
+import { configuration, OutputLevel } from './configuration';
+import { ExtensionOutputChannelName } from './constants';
 
 const ConsolePrefix = `[${ExtensionOutputChannelName}]`;
 
-export enum OutputLevel {
-    Silent = 'silent',
-    Errors = 'errors',
-    Verbose = 'verbose'
-}
-
-let debug = false;
-let level: OutputLevel = OutputLevel.Silent;
-let output: OutputChannel;
-
-function onConfigurationChanged() {
-    const cfg = workspace.getConfiguration().get<IConfig>(ExtensionKey);
-    if (cfg === undefined) return;
-
-    if (cfg.debug !== debug || cfg.outputLevel !== level) {
-        debug = cfg.debug;
-        level = cfg.outputLevel;
-
-        if (level === OutputLevel.Silent) {
-            output && output.dispose();
-        }
-        else {
-            output = output || window.createOutputChannel(ExtensionOutputChannelName);
-        }
-    }
-}
-
 export class Logger {
 
+    static debug = false;
+    static level: OutputLevel = OutputLevel.Silent;
+    static output: OutputChannel | undefined;
+
     static configure(context: ExtensionContext) {
-        context.subscriptions.push(workspace.onDidChangeConfiguration(onConfigurationChanged));
-        onConfigurationChanged();
+        context.subscriptions.push(configuration.onDidChange(this.onConfigurationChanged, this));
+        this.onConfigurationChanged(configuration.initializingChangeEvent);
+    }
+
+    private static onConfigurationChanged(e: ConfigurationChangeEvent) {
+        const initializing = configuration.initializing(e);
+
+        let section = configuration.name('debug').value;
+        if (initializing || configuration.changed(e, section)) {
+            this.debug = configuration.get<boolean>(section);
+        }
+
+        section = configuration.name('outputLevel').value;
+        if (initializing || configuration.changed(e, section)) {
+            this.level = configuration.get<OutputLevel>(section);
+
+            if (this.level === OutputLevel.Silent) {
+                if (this.output !== undefined) {
+                    this.output.dispose();
+                    this.output = undefined;
+                }
+            }
+            else {
+                this.output = this.output || window.createOutputChannel(ExtensionOutputChannelName);
+            }
+        }
     }
 
     static log(message?: any, ...params: any[]): void {
-        if (debug) {
-            console.log(ConsolePrefix, message, ...params);
+        if (this.debug) {
+            console.log(this.timestamp, ConsolePrefix, message, ...params);
         }
 
-        if (level === OutputLevel.Verbose) {
-            output.appendLine([message, ...params].join(' '));
+        if (this.output !== undefined && (this.level === OutputLevel.Verbose || this.level === OutputLevel.Debug)) {
+            this.output.appendLine((this.debug ? [this.timestamp, message, ...params] : [message, ...params]).join(' '));
         }
     }
 
     static error(ex: Error, classOrMethod?: string, ...params: any[]): void {
-        if (debug) {
-            console.error(ConsolePrefix, classOrMethod, ex, ...params);
+        if (this.debug) {
+            console.error(this.timestamp, ConsolePrefix, classOrMethod, ...params, ex);
         }
 
-        if (level !== OutputLevel.Silent) {
-            output.appendLine([classOrMethod, ex, ...params].join(' '));
+        if (this.output !== undefined && this.level !== OutputLevel.Silent) {
+            this.output.appendLine((this.debug ? [this.timestamp, classOrMethod, ...params, ex] : [classOrMethod, ...params, ex]).join(' '));
         }
     }
 
     static warn(message?: any, ...params: any[]): void {
-        if (debug) {
-            console.warn(ConsolePrefix, message, ...params);
+        if (this.debug) {
+            console.warn(this.timestamp, ConsolePrefix, message, ...params);
         }
 
-        if (level !== OutputLevel.Silent) {
-            output.appendLine([message, ...params].join(' '));
+        if (this.output !== undefined && this.level !== OutputLevel.Silent) {
+            this.output.appendLine((this.debug ? [this.timestamp, message, ...params] : [message, ...params]).join(' '));
         }
+    }
+
+    private static get timestamp(): string {
+        const now = new Date();
+        return `[${now.toISOString().replace(/T/, ' ').replace(/\..+/, '')}:${('00' + now.getUTCMilliseconds()).slice(-3)}]`;
     }
 }