+++ src/system.ts
@@ -1,3 +1,6 @@
 'use strict';
+
 export * from './system/array';
+export * from './system/function';
 export * from './system/object';
+export * from './system/string';